migrate the wallet page from /wallet to /app for shortened URL on up.io
